requirements.txt
ccxt>=4.0.0
aiohttp>=3.9.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
aiosqlite>=0.19.0
